some fixies: 1. moved classes to unother dir 2/ creted BaseEntity 3. added extends BaseEntity to Entity Classes
