+++ src/main/java/ru/zhelper/zhelper/models/BaseStatus.java
@@ -1,0 +1,5 @@
+package ru.zhelper.zhelper.models;
+
+public enum BaseStatus {
+    ACTIVE, NOT_ACTIVE, DELETED, BANNED
+}
