+++ src/main/java/ru/zhelper/zhelper/models/users/Role.java
@@ -1,16 +1,14 @@
 package ru.zhelper.zhelper.models.users;
 
 import lombok.Data;
+import ru.zhelper.zhelper.models.BaseEntity;
 
 import javax.persistence.*;
 
 @Entity
 @Table(name = "roles")
 @Data
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+public class Role extends BaseEntity {
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
