+++ src/main/java/ru/zhelper/zhelper/models/users/User.java
@@ -2,12 +2,12 @@ package ru.zhelper.zhelper.models.users;
 
 import lombok.Data;
 import org.hibernate.annotations.Type;
+import ru.zhelper.zhelper.models.BaseEntity;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
-import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -19,11 +19,7 @@ import java.util.Set;
                 @UniqueConstraint(columnNames = "email"),
                 @UniqueConstraint(columnNames = "telegramUserId")
         })
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+public class User extends BaseEntity {
     @NotBlank(message = "Name is mandatory")
     @Size(max = 30)
     private String username;
@@ -43,16 +39,8 @@ public class User {
             inverseJoinColumns = @JoinColumn(name = "role_id"))
     private Set<Role> roles = new HashSet<>();
 
-    private LocalDateTime created;
-
-    private LocalDateTime edited;
-
     @Lob
     @Column(length = 10000)
     @Type(type = "org.hibernate.type.TextType")
     private String comment;
-
-    public User() {
-        this.created = LocalDateTime.now();
-    }
 }
