+++ src/main/java/ru/zhelper/zhelper/repository/ProcurementRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import ru.zhelper.zhelper.models.Procurement;
+import ru.zhelper.zhelper.models.procurements.Procurement;
 
 import java.time.LocalDate;
 import java.util.List;
