+++ src/main/java/ru/zhelper/zhelper/services/ProcurementServiceImpl.java
@@ -2,7 +2,7 @@ package ru.zhelper.zhelper.services;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
-import ru.zhelper.zhelper.models.Procurement;
+import ru.zhelper.zhelper.models.procurements.Procurement;
 import ru.zhelper.zhelper.models.dto.ProcurementAddress;
 import ru.zhelper.zhelper.repository.ProcurementRepo;
 import ru.zhelper.zhelper.services.parsers_dispatcher.Dispatcher;
