+++ src/main/java/ru/zhelper/zhelper/services/chrome/ProcurementDtoServiceImpl.java
@@ -3,8 +3,8 @@ package ru.zhelper.zhelper.services.chrome;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import ru.zhelper.zhelper.models.ProcedureType;
-import ru.zhelper.zhelper.models.Procurement;
+import ru.zhelper.zhelper.models.procurements.ProcedureType;
+import ru.zhelper.zhelper.models.procurements.Procurement;
 import ru.zhelper.zhelper.models.dto.ProcurementDto;
 import ru.zhelper.zhelper.services.dao.ProcurementDataManager;
 import ru.zhelper.zhelper.services.exceptions.BadDataParsingException;
