+++ src/main/java/ru/zhelper/zhelper/services/parsers_dispatcher/CentralDispatcher.java
@@ -3,8 +3,8 @@ package ru.zhelper.zhelper.services.parsers_dispatcher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import ru.zhelper.zhelper.models.Procurement;
-import ru.zhelper.zhelper.models.ProcurementType;
+import ru.zhelper.zhelper.models.procurements.Procurement;
+import ru.zhelper.zhelper.models.procurements.ProcurementType;
 import ru.zhelper.zhelper.services.exceptions.BadDataParsingException;
 import ru.zhelper.zhelper.services.exceptions.BadRequest;
 import ru.zhelper.zhelper.services.parsers_dispatcher.parser.ZakupkiParser;
