+++ src/main/java/ru/zhelper/zhelper/services/parsers_dispatcher/Dispatcher.java
@@ -1,6 +1,6 @@
 package ru.zhelper.zhelper.services.parsers_dispatcher;
 
-import ru.zhelper.zhelper.models.Procurement;
+import ru.zhelper.zhelper.models.procurements.Procurement;
 
 public interface Dispatcher {
     Procurement getFromUrl(String url);
