+++ src/main/java/ru/zhelper/zhelper/services/parsers_dispatcher/parser/ZakupkiParser.java
@@ -1,6 +1,6 @@
 package ru.zhelper.zhelper.services.parsers_dispatcher.parser;
 
-import ru.zhelper.zhelper.models.Procurement;
+import ru.zhelper.zhelper.models.procurements.Procurement;
 
 public interface ZakupkiParser {
     Procurement parse(String html);
