+++ src/main/java/ru/zhelper/zhelper/services/validator/URLValidator.java
@@ -1,6 +1,6 @@
 package ru.zhelper.zhelper.services.validator;
 
-import ru.zhelper.zhelper.models.ProcurementType;
+import ru.zhelper.zhelper.models.procurements.ProcurementType;
 
 public interface URLValidator {
     boolean isValidUrl (String url);
