+++ src/main/java/ru/zhelper/zhelper/services/validator/URLValidatorImpl.java
@@ -1,7 +1,7 @@
 package ru.zhelper.zhelper.services.validator;
 
 import org.springframework.stereotype.Service;
-import ru.zhelper.zhelper.models.ProcurementType;
+import ru.zhelper.zhelper.models.procurements.ProcurementType;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
