+++ src/test/java/ru/zhelper/zhelper/services/ProcurementDataManagerImplTest.java
@@ -17,9 +17,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
-import ru.zhelper.zhelper.models.ProcedureType;
-import ru.zhelper.zhelper.models.Procurement;
-import ru.zhelper.zhelper.models.Stage;
+import ru.zhelper.zhelper.models.procurements.ProcedureType;
+import ru.zhelper.zhelper.models.procurements.Procurement;
+import ru.zhelper.zhelper.models.procurements.Stage;
 import ru.zhelper.zhelper.repository.ProcurementRepo;
 import ru.zhelper.zhelper.services.dao.ProcurementDataManagerImpl;
 import ru.zhelper.zhelper.services.exceptions.DataManagerException;
