+++ src/test/java/ru/zhelper/zhelper/services/parsers_dispatcher/CentralDispatcherTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import ru.zhelper.zhelper.models.Procurement;
-import ru.zhelper.zhelper.models.ProcurementType;
+import ru.zhelper.zhelper.models.procurements.Procurement;
+import ru.zhelper.zhelper.models.procurements.ProcurementType;
 import ru.zhelper.zhelper.services.exceptions.BadDataParsingException;
 import ru.zhelper.zhelper.services.exceptions.BadRequest;
 import ru.zhelper.zhelper.services.parsers_dispatcher.parser.ZakupkiParser;
